refactor(p1am): move pass-through repair helper
